Moved all optimization steps from codeopt.c and some optimization steps from coptind.c into new separate files.
