+++ src/cc65.vcxproj
@@ -93,6 +93,8 @@
     <ClInclude Include="cc65\coptc02.h" />
     <ClInclude Include="cc65\coptcmp.h" />
     <ClInclude Include="cc65\coptind.h" />
+    <ClInclude Include="cc65\coptjmp.h" />
+    <ClInclude Include="cc65\coptmisc.h" />
     <ClInclude Include="cc65\coptneg.h" />
     <ClInclude Include="cc65\coptptrload.h" />
     <ClInclude Include="cc65\coptptrstore.h" />
@@ -167,6 +169,8 @@
     <ClCompile Include="cc65\coptc02.c" />
     <ClCompile Include="cc65\coptcmp.c" />
     <ClCompile Include="cc65\coptind.c" />
+    <ClCompile Include="cc65\coptjmp.c" />
+    <ClCompile Include="cc65\coptmisc.c" />
     <ClCompile Include="cc65\coptneg.c" />
     <ClCompile Include="cc65\coptptrload.c" />
     <ClCompile Include="cc65\coptptrstore.c" />
